+++ format-importer.ts
@@ -1,0 +1,12 @@
+import { App } from "obsidian";
+import { ImportResult } from "yarle/interfaces";
+
+export abstract class FormatImporter {
+	app: App;
+
+	constructor(app: App) {
+		this.app = app;
+	}
+
+        abstract import(filePaths: string[], outputFolder: string): Promise<ImportResult>;
+}
