+++ yarle/interfaces.ts
@@ -1,0 +1,5 @@
+export interface ImportResult {
+	total: number,
+	failed: number,
+	skipped: number
+}
